Update the solver user code interface commands so that they take an argument. THey can be called from general user code now instead of just from the solver.
They use five registers and two or three flags:

Registers
       n+0: first estimate a
       n+1: second estimate b
       n+2: third estimate c
       n+3: function value at first estimate f(a)
       n+4: function value at second estimate f(b)

Flags
       n+1: true if bracketing solution - one guess positive and the other negative
       n+2: true if the function appears to be constant
       n+3: true if a bisection step was done last (only used if Ridder's step is included -- it isn't currently)


SLVI takes values for a, b, f(a) and f(b) in these registers and
produces a guess returned in register c.  There is no stack
interaction.

SLVS takes values in all five registers plus f(c) in X and updates
all the register values so that c and f(c) replace one of a & f(a) or
b & f(b).  It also produces a new guess returned in register c and
returns 0 in X if the solving should continue and non-zero if not.
Otherwise, the stack isn't altered.


The built in solver loop looks kind of like:

==============================================
       calculate f(a) & f(b) and iniitalise the registers

	SLVI
	LBL 00
		RCL 02		; register C
		XEQ 02		; call the user's subroutine
		x[approx]0?	; test if the solution has converged
		GTO 01		; converged, so exit
		SLVS		; update estimates
		x=0?		; should we continue?
		GTO 00		; loop back again
	LBL 01
	RCL 02			; best guess so far
	RTN
==============================================

This isn't quite it since I fill the stack and test if the two guesses
a & b are approximately equal but it gives the general idea of the
effort required.


The actual solver is fairly complex.  A combination of quadratic
interpolation, a guarded secant method, bisection and Ridder's step
are used (the latter being optional and disabled currently).

The Ridder's step is still uncertain.  I'm not convinced it gives us
improved convergence over using just the other methods.  At least not
enough of a benefit to justify the space required.



git-svn-id: http://svn.code.sf.net/p/wp34s/code@67 9654ca62-1852-461c-aa49-1be32f3bb02d