Update the quadratic solver to use numerically stable algorithms and handle complex roots.
Input is a in Z, b in Y and c in X.

The equation 0 = a x^2 + b x + c is solved.

For real roots, the solutions are in X and Y.

For complex roots, one root is in X and Y the other root is the conjugate
of X and Y.  In a program the step after SLVQ is skipped, if run directly
by a user, the complex result 'C' is displayed.


In either case, b^2 - 4 a c is in Z.  T is unchanged.  L has the value of c.




git-svn-id: http://svn.code.sf.net/p/wp34s/code@1454 9654ca62-1852-461c-aa49-1be32f3bb02d